requirements.txt
chromadb
sentence-transformers
python-dotenv
flask
flask-cors
tiktoken
openai
watchdog
nltk
requests
